Add PR changes and a .d.ts file for types
